+++ app/views/actions/projects.js
@@ -187,9 +187,9 @@ export function saveNote(parentsId, uuid) {
   };
 }
 
-export const UPLOAD_NOTE_ONEDRIVER = 'UPLOAD_NOTE_ONEDRIVER';
-export const UPLOAD_NOTE_ONEDRIVER_SUCCESS = 'UPLOAD_NOTE_ONEDRIVER_SUCCESS';
-export const UPLOAD_NOTE_ONEDRIVER_FAILED = 'UPLOAD_NOTE_ONEDRIVER_FAILED';
+export const UPLOAD_NOTE_ONEDRIVE = 'UPLOAD_NOTE_ONEDRIVE';
+export const UPLOAD_NOTE_ONEDRIVE_SUCCESS = 'UPLOAD_NOTE_ONEDRIVE_SUCCESS';
+export const UPLOAD_NOTE_ONEDRIVE_FAILED = 'UPLOAD_NOTE_ONEDRIVE_FAILED';
 
 export const UPDATE_NOTE_UPLOAD_STATUS = 'UPDATE_NOTE_UPLOAD_STATUS';
 export function updateNoteUploadStatus(parentsId, uuid, status) {
@@ -201,4 +201,4 @@ export function updateNoteUploadStatus(parentsId, uuid, status) {
   };
 }
 
-export const SAVE_NOTE_FROM_DRIVER = 'SAVE_NOTE_FROM_DRIVER';
+export const SAVE_NOTE_FROM_DRIVE = 'SAVE_NOTE_FROM_DRIVE';