+++ app/views/reducers/projects.js
@@ -18,11 +18,11 @@ import {
   TRASH_CHOOSE_PROJECT,
   TRASH_BACK_ROOT,
   SAVE_NOTE_ON_KEYDOWN,
-  UPLOAD_NOTE_ONEDRIVER,
-  UPLOAD_NOTE_ONEDRIVER_SUCCESS,
-  UPLOAD_NOTE_ONEDRIVER_FAILED,
+  UPLOAD_NOTE_ONEDRIVE,
+  UPLOAD_NOTE_ONEDRIVE_SUCCESS,
+  UPLOAD_NOTE_ONEDRIVE_FAILED,
   UPDATE_NOTE_UPLOAD_STATUS,
-  SAVE_NOTE_FROM_DRIVER,
+  SAVE_NOTE_FROM_DRIVE,
 } from '../actions/projects';
 import {
   getProjectList,
@@ -735,7 +735,7 @@ function projectReducer(state = {
       updateNoteInfo(uuid, param);
       return assign({}, state);
     }
-    case UPLOAD_NOTE_ONEDRIVER: {
+    case UPLOAD_NOTE_ONEDRIVE: {
       const { param: { uuid, projectUuid } } = action;
       const notes = state.projects[state.hash[projectUuid]].notes.map((note) => {
         if (note.uuid === uuid) {
@@ -747,7 +747,7 @@ function projectReducer(state = {
       dbUpdateNote({ uuid }, { oneDriver: 2 });
       return assign({}, state);
     }
-    case UPLOAD_NOTE_ONEDRIVER_SUCCESS: { // 单个笔记上传成功
+    case UPLOAD_NOTE_ONEDRIVE_SUCCESS: { // 单个笔记上传成功
       const { param: { uuid, projectUuid } } = action;
       const notes = state.projects[state.hash[projectUuid]].notes.map((note) => {
         if (note.uuid === uuid) {
@@ -759,7 +759,7 @@ function projectReducer(state = {
       dbUpdateNote({ uuid }, { oneDriver: 3 });
       return assign({}, state);
     }
-    case UPLOAD_NOTE_ONEDRIVER_FAILED: {
+    case UPLOAD_NOTE_ONEDRIVE_FAILED: {
       const { param: { uuid, projectUuid } } = action;
       const notes = state.projects[state.hash[projectUuid]].notes.map((note) => {
         if (note.uuid === uuid) {
@@ -782,8 +782,8 @@ function projectReducer(state = {
       state.projects[state.hash[parentsId]].notes = notes;
       return assign({}, state);
     }
-    case SAVE_NOTE_FROM_DRIVER: { // 保存从云端下载的笔记
-      const { folder, name, content, info, driverType } = action;
+    case SAVE_NOTE_FROM_DRIVE: { // 保存从云端下载的笔记
+      const { folder, name, content, info, driveType } = action;
       const { projects } = state;
       let needCreateProject = true;
       let targetProject;
@@ -806,7 +806,7 @@ function projectReducer(state = {
           labels: info.labels,
           description: info.description,
           createDate,
-          [driverType]: 3,
+          [driveType]: 3,
         };
         const file = createFile(fileInfo);
         newProject.notes = [file];
@@ -834,7 +834,7 @@ function projectReducer(state = {
             labels: info.labels,
             description: info.description,
             createDate,
-            [driverType]: 3,
+            [driveType]: 3,
           };
           const file = createFile(fileInfo);
           targetProject.notes.unshift(file);