+++ app/views/sagas/app.js
@@ -2,36 +2,36 @@ import { put, call, takeLatest } from 'redux-saga/effects';
 import { message } from 'antd';
 // import { ipcRenderer } from 'electron';
 import {
-  FETCHING_ONEDRIVER_TOKEN,
-  FETCHING_ONEDRIVER_TOKEN_FAILED,
-  FETCHING_ONEDRIVER_TOKEN_SUCCESS,
+  FETCHING_ONEDRIVE_TOKEN,
+  FETCHING_ONEDRIVE_TOKEN_FAILED,
+  FETCHING_ONEDRIVE_TOKEN_SUCCESS,
   FETCHING_GITHUB_RELEASES,
   FETCHING_GITHUB_RELEASES_FAILED,
   FETCHING_GITHUB_RELEASES_SUCCESS,
 } from '../actions/app';
 
-import OneDriver from '../services/OneDriver';
+import OneDrive from '../services/OneDrive';
 import CommonServices from '../services/CommonServices';
 
-const oneDriver = new OneDriver();
+const oneDrive = new OneDrive();
 const commonServices = new CommonServices();
 
-function* oneDriverToken(action) {
+function* oneDriveToken(action) {
   const { code } = action;
   try {
-    const data = yield call(oneDriver.getTokenByCode, code);
+    const data = yield call(oneDrive.getTokenByCode, code);
     const token = data.access_token;
     const refreshToken = data.refresh_token;
     const expiresDate = Date.parse(new Date()) + (data.expires_in * 1000);
-    yield put({ type: FETCHING_ONEDRIVER_TOKEN_SUCCESS, token, refreshToken, expiresDate });
+    yield put({ type: FETCHING_ONEDRIVE_TOKEN_SUCCESS, token, refreshToken, expiresDate });
   } catch (ex) {
     message.error('Auth failed');
-    yield put({ type: FETCHING_ONEDRIVER_TOKEN_FAILED, error: ex });
+    yield put({ type: FETCHING_ONEDRIVE_TOKEN_FAILED, error: ex });
   }
 }
 
-function* fetchingOneDriverToken() {
-  yield takeLatest(FETCHING_ONEDRIVER_TOKEN, oneDriverToken);
+function* fetchingOneDriveToken() {
+  yield takeLatest(FETCHING_ONEDRIVE_TOKEN, oneDriveToken);
 }
 
 function* handleReleaseFetch() {
@@ -48,6 +48,6 @@ function* getReleases() {
 }
 
 export default [
-  fetchingOneDriverToken,
+  fetchingOneDriveToken,
   getReleases,
 ];